# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation - Basic CRUD Laravel",
        "contact": {
            "email": "manirujjamanakash@gmail.com"
        },
        "license": {
            "name": "@ManiruzzamanAkash",
            "url": "https://devsenv.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "manirujjamanakash@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticated User Profile",
                "description": "Authenticated User Profile",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::me",
                "responses": {
                    "200": {
                        "description": "Authenticated User Profile"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Logout"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh",
                "description": "Refresh",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::refresh",
                "responses": {
                    "200": {
                        "description": "Refresh"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product List",
                "description": "Get Product List as Array",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Get Product List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create New Product",
                "description": "Create New Product",
                "operationId": "App\\Http\\Controllers\\Products\\ProductsController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Product 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10120
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Product Details",
                "description": "Show Product Details",
                "operationId": "App\\Http\\Controllers\\Products\\ProductsController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Product Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Update Product",
                "operationId": "App\\Http\\Controllers\\Products\\ProductsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Product 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10120
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "description": "Delete Product",
                "operationId": "App\\Http\\Controllers\\Products\\ProductsController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/": {
            "get": {
                "description": "Home page",
                "responses": {
                    "default": {
                        "description": "Welcome page"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}